+++ dictionary/get_genc.R
@@ -1,11 +1,14 @@
+source(here::here("dictionary/utilities.R"))
+
 # Check https://geonames.nga.mil/gns/html/countrycodes.html quarterly for the latest XLSX url
 url <- 'http://geonames.nga.mil/gns/html/docs/GENC_ED3U12_GEC_XWALK.xlsx'
 
 # Uses DoD Root CA 3, which I don't have in my root certificate store
 httr::set_config(config(ssl_verifypeer = 0L))
 
-httr::GET(url, write_disk(filename, overwrite=TRUE))
-genc <- readxl::read_excel(filename, sheet = 'Codes_for_GE_Names', skip = 2)
+tmp <- tempfile()
+httr::GET(url, write_disk(tmp, overwrite=TRUE))
+genc <- readxl::read_excel(tmp, sheet = 'Codes_for_GE_Names', skip = 2)
 
 bad <- c('AKROTIRI', 'ASHMORE AND CARTIER ISLANDS', 'BAKER ISLAND',
          'CLIPPERTON ISLAND', 'CORAL SEA ISLANDS', 'DHEKELIA', 'DIEGO GARCIA',
@@ -21,4 +24,8 @@ genc <-
                   genc3c = `3-character Code`, genc3n = `Numeric Code`, ) %>%
     dplyr::filter(!country %in% bad)
 
-genc %>% write_csv('dictionary/data_genc.csv')
+# sanity checks
+checkmate::assert_true("NA" == genc$genc2c[tolower(genc$country) == "namibia"])
+
+# write
+genc %>% write_csv('dictionary/data_genc.csv', na = "")