+++ dictionary/get_world_bank_api.R
@@ -8,4 +8,4 @@ wb_api <- fromJSON(tmp, flatten = T)[[2]] %>%
           filter(region.value != 'Aggregates') %>%
           select(country = name, wb_api2c = iso2Code, wb_api3c = id)
 
-wb_api %>% write_csv('dictionary/data_world_bank_api.csv')
+wb_api %>% write_csv('dictionary/data_world_bank_api.csv', na = "")