+++ NEWS.md
@@ -1,5 +1,6 @@
 # countrycode 1.2.0.9000
 
+* fixed codes for Namibia: eurostat, genc2c, wb_api2c, ecb
 * regex improvements
 * countryname(warn=TRUE) by default
 * better class checks